+++ src/components/Footer/Footer.stories.tsx
@@ -9,12 +9,12 @@ import Instagram from '../../assets/instagram.svg?react';
 import Wechat from '../../assets/wechat.svg?react';
 import Spotify from '../../assets/spotify.svg?react';
 import Social from '../Social/Social';
+import { SubscriptionState } from '../Subscribe/types';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
   title: 'Components/Footer',
   component: Footer,
-  tags: ['autodocs'],
 } satisfies Meta<typeof Footer>;
 
 export default meta;
@@ -76,15 +76,19 @@ const leftComponent = (
     element="form"
     buttonProps={{
       size: 'sm',
-      onClick: (e: React.MouseEvent) => {
-        e.preventDefault();
-        const inputElement = (e.target as HTMLElement).closest('form')?.querySelector('input');
-        if (inputElement) {
-          console.log(`subscribe ${inputElement.value}`);
-          inputElement.value = '';
-        }
-      },
     }}
+    onSubmit={(e: React.FormEvent<HTMLFormElement>) => {
+      const target = e.target as HTMLFormElement;
+      e.preventDefault();
+
+      const nativeData = new FormData(target);
+      const data = Object.fromEntries(nativeData.entries());
+      console.log('Form submitted for email -', data?.email);
+
+      target.reset();
+    }}
+    subscriptionState={SubscriptionState.Default}
+    invalidText="Please enter a valid email address."
   />
 );