+++ src/components/GridItem/GridItem.stories.tsx
@@ -6,7 +6,6 @@ import { GridItemAlign } from './types';
 const meta = {
   title: 'Components/Layouts/GridItem',
   component: GridItem,
-  tags: ['autodocs'],
 } satisfies Meta<typeof GridItem>;
 
 export default meta;
@@ -22,9 +21,18 @@ Playground.args = {
   children: <div>Grid Item</div>,
 };
 
-export const CenteredGridItem = () => (
+Playground.argTypes = {
+  align: {
+    options: GridItemAlign,
+    control: {
+      type: 'select',
+    },
+  },
+};
+
+export const CenteredGridItem = (props: GridItemProps) => (
   <Grid>
-    <GridItem style={{ backgroundColor: 'gray' }} sm={2} md={8} lg={8} align={GridItemAlign.center}>
+    <GridItem style={{ backgroundColor: 'gray' }} {...props} align={GridItemAlign.center}>
       <p>
         Harry Phillips founded the auction house in 1796 in Westminster, London. Phillips gained international
         recognition by selling paintings from the estate of Queen Marie Antoinette and household items from Napoleon
@@ -33,7 +41,7 @@ export const CenteredGridItem = () => (
         receptions before auctions – a standard practice in the auction business today.
       </p>
     </GridItem>
-    <GridItem style={{ backgroundColor: 'gray' }} sm={2} md={8} lg={8} align={GridItemAlign.center}>
+    <GridItem style={{ backgroundColor: 'gray' }} {...props} align={GridItemAlign.center}>
       <p>
         Harry Phillips founded the auction house in 1796 in Westminster, London. Phillips gained international
         recognition by selling paintings from the estate of Queen Marie Antoinette and household items from Napoleon
@@ -44,6 +52,12 @@ export const CenteredGridItem = () => (
     </GridItem>
   </Grid>
 );
+
+CenteredGridItem.args = {
+  sm: 2,
+  md: 8,
+  lg: 8,
+};
 
 export const LeftAndRightGridItems = () => (
   <Grid>