+++ src/components/Link/Link.stories.tsx
@@ -6,7 +6,6 @@ import { LinkVariants } from './utils';
 const meta = {
   title: 'Components/Links/Link',
   component: Link,
-  tags: ['autodocs'],
 } satisfies Meta<typeof Link>;
 
 export default meta;
@@ -27,6 +26,15 @@ Playground.args = {
   children: 'My Link',
   href: internalHref,
   variant: LinkVariants.standalone,
+};
+
+Playground.argTypes = {
+  variant: {
+    options: LinkVariants,
+    control: {
+      type: 'select',
+    },
+  },
 };
 
 export const CustomLink = ({ playgroundWidth, children, ...args }: StoryProps) => (