+++ src/components/Select/Select.stories.tsx
@@ -5,7 +5,6 @@ import Select, { SelectProps } from './Select';
 const meta = {
   title: 'Components/Select',
   component: Select,
-  tags: ['autodocs'],
 } satisfies Meta<typeof Select>;
 
 export default meta;