+++ src/components/Subscribe/Subscribe.tsx
@@ -3,6 +3,7 @@ import classnames from 'classnames';
 import { px } from '../../utils';
 import Input from '../Input/Input';
 import Button, { ButtonProps } from '../Button/Button';
+import { SubscriptionState } from './types';
 
 export interface SubscribeProps extends React.HTMLAttributes<HTMLFormElement> {
   /**
@@ -33,6 +34,22 @@ export interface SubscribeProps extends React.HTMLAttributes<HTMLFormElement> {
    * Subscribe title text
    */
   title?: string;
+  /**
+   * Subscribe loading text
+   */
+  loadingText?: string;
+  /**
+   * Subscribe error text
+   */
+  invalidText?: string;
+  /**
+   * Subscribe success text
+   */
+  successText?: string;
+  /**
+   * Subscribe state for loading or error
+   */
+  subscriptionState?: SubscriptionState;
 }
 
 /**
@@ -52,22 +69,47 @@ const Subscribe = ({
   inputLabelText = 'Email*',
   inputPlaceholder = 'example@email.com',
   title = 'Subscribe to Newsletter',
+  loadingText = 'Loading...',
+  invalidText = '',
+  successText,
+  subscriptionState = SubscriptionState.Default,
   ...props
 }: SubscribeProps) => {
+  const isInvalid = subscriptionState === 'invalid';
+  const isLoading = subscriptionState === 'loading';
+  const isSuccess = subscriptionState === 'success';
+
+  const textsCondition = {
+    invalid: invalidText,
+    success: successText,
+    loading: loadingText,
+  };
+
+  const warnText = subscriptionState !== 'default' ? textsCondition[subscriptionState] : '';
+  const warn = isSuccess || isLoading;
+
   return (
     <Element
       data-testid={id ? id : `subscribe-form`}
       id={id}
       className={classnames(`${px}-subscribe`, className)}
+      noValidate
       {...props}
     >
       <h3 className={`${px}-subscribe__title`}>{title}</h3>
       {blurb ? <p className={`${px}-subscribe__blurb`}>{blurb}</p> : null}
+
       <Input
         className={`${px}-subscribe__input`}
         type="email"
+        name="email"
         placeholder={inputPlaceholder}
         labelText={inputLabelText}
+        invalid={isInvalid}
+        invalidText={invalidText}
+        warn={warn}
+        warnText={warnText}
+        required
       />
       <Button className={`${px}-subscribe__button ${className}`} buttonType="secondary" type="submit" {...buttonProps}>
         {buttonText}