+++ src/index.ts
@@ -23,5 +23,10 @@ export {
   default as StatefulViewingsList,
   type StatefulViewingsListProps,
 } from './components/ViewingsList/StatefulViewingsList';
+export * from './components/Text';
+
+// Utils
+export { PaddingTokens } from './utils';
+
 // 📑 Pages
 export { default as Page } from './pages/Page';